pyproject.toml
[project]
name = "budget-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "faker>=39.0.0",
    "fastapi[standard]>=0.127.0",
    "pandas>=2.3.3",
    "python-telegram-bot>=22.5",
]
